feat(pose_initializer): move helpers
