+++ localization/pose_initializer/CMakeLists.txt
@@ -13,20 +13,11 @@ ament_auto_add_executable(pose_initializer_node
   src/pose_initializer/ndt_module.cpp
   src/pose_initializer/stop_check_module.cpp
 )
-
 ament_auto_add_executable(map_height_fitter
   src/map_height_fitter/map_height_fitter_node.cpp
   src/map_height_fitter/map_height_fitter_core.cpp
 )
 target_link_libraries(map_height_fitter ${PCL_LIBRARIES})
-
-ament_auto_add_library(${PROJECT_NAME} SHARED
-  src/lib/initial_pose_adaptor.cpp
-  src/lib/automatic_api_caller.cpp
-)
-
-ament_auto_add_executable(initial_pose_adaptor src/initial_pose_adaptor.cpp)
-ament_auto_add_executable(automatic_api_caller src/automatic_api_caller.cpp)
 
 if(BUILD_TESTING)
   function(add_testcase filepath)
