+++ localization/pose_initializer/config/pose_initializer.param.yaml
@@ -1,17 +1,6 @@
 /**:
   ros__parameters:
 
-    # from initialpose (Rviz's 2DPoseEstimate)
-    initialpose_particle_covariance:
-      [
-        4.0, 0.0, 0.0,  0.0,  0.0,  0.0,
-        0.0, 4.0, 0.0,  0.0,  0.0,  0.0,
-        0.0, 0.0, 0.01, 0.0,  0.0,  0.0,
-        0.0, 0.0, 0.0,  0.01, 0.0,  0.0,
-        0.0, 0.0, 0.0,  0.0,  0.01, 0.0,
-        0.0, 0.0, 0.0,  0.0,  0.0,  1.0,
-      ]
-
     # from gnss
     gnss_particle_covariance:
       [
