+++ system/default_ad_api_helpers/config/initial_pose.param.yaml
@@ -1,0 +1,13 @@
+/**:
+  ros__parameters:
+
+    # from initialpose (Rviz's 2DPoseEstimate)
+    initial_pose_particle_covariance:
+      [
+        4.0, 0.0, 0.0,  0.0,  0.0,  0.0,
+        0.0, 4.0, 0.0,  0.0,  0.0,  0.0,
+        0.0, 0.0, 0.01, 0.0,  0.0,  0.0,
+        0.0, 0.0, 0.0,  0.01, 0.0,  0.0,
+        0.0, 0.0, 0.0,  0.0,  0.01, 0.0,
+        0.0, 0.0, 0.0,  0.0,  0.0,  1.0,
+      ]
