+++ system/default_ad_api_helpers/src/automatic_api_caller.cpp
